move sstable_index to versioned file name
